requirements.txt
numpy<2.0.0
pandas
mplfinance
yfinance
TA-Lib
pyyaml
ray[default]
scipy

#--- OPTIONAL ---
bayesian-optimization